Sticky cell interface
There's been a long standing demand in the Jupyter notebook for
sticky cell support.  That is, users want to be able to pin
cells somewhere within view so they remain visible while scrolling
through the document.

This mock suggests a pin UI button be added to the cell toolbar.

![pin cell](sticky.jpg)

When a cell is pinned, it will appear in a pane that floats
independent of the scroll position of the document.

![pinned cell](stickied.jpg)

Although the mock above shows the hover region vertically above
the document, I think there's reason to investigate it being to
the side of the document instead.  Benifits to a side by side
layout include the ability for the floating area to occupy more
vertical space.